fix optionselection for getopt
Wow, this is embarrassing:
When using the short option "-k" for "--keepalive", getopt (due to
missing ":" after "k") will not know to expect a parameter.
Thus the application will segfault and crash when parsing the
short-option.
However the long-option works (and this is most probably how
it passed then tests).

This patch fixes the issue by correctly announcing (via ":") the
short-option "k" must have a paramter.

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>